backends/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.4.1
Django==1.11.16
django-cors-headers==2.4.0
django-extensions==2.1.3
django-redis==4.9.0
enum==0.4.7
enum34==1.1.6
functools32==3.2.3.post2
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
jsonschema==2.6.0
mysqlclient==1.3.13
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.6.4
pytz==2018.7
redis==2.10.6
six==1.11.0
typing==3.6.6
